% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aski.R
\name{aski}
\alias{aski}
\title{Aski}
\usage{
aski()
}
\arguments{
\item{input}{Optional: A string containing the error or question. If not provided, the clipboard content is used.}
}
\value{
A recommendation for addressing the error.
}
\description{
Obtain AI assistance for resolving error messages.
Simply copy your error, execute aski(), and watch your error get resolved in no time!
}
\details{
By default, aski accesses your clipboard - No input needed.
Alternatively, you can enter string containing the error or question.
}
