% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setup.R
\name{check_setup}
\alias{check_setup}
\title{Check Setup}
\usage{
check_setup()
}
\description{
Determines if the OpenAI API Key has been configured.

If you don't already have an API key, you can obtain one by signing up at https://platform.openai.com/signup.
}
\examples{
\dontrun{
# Check if the API key is configured
check_setup()
}#'
}
