% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_key}
\alias{set_key}
\title{Set key}
\usage{
set_key(key)
}
\arguments{
\item{key}{Character string containing the OpenAI API key.}
}
\description{
This function sets the OpenAI API key and stores it for later use. It also sets the key as an environment variable for the current R session.
}
\examples{
set_key("abc123def456")



}
